+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -15,6 +15,8 @@
 - `ExpiryTime` is renamed to `ExpiresOn`.
 - `LastAccessTime` is renamed to `LastAccessedOn`.
 - Move version strings into `Details` namespace.
+- `ReadFileResult` now have `ContentRange` as string.
+- `ReadFileOptions` now have `Azure::Core::Http::Range Range` instead of `Content-Length` and `Offset`.
 
 ## 12.0.0-beta.5 (2020-11-13)